Updated Devnet Client
