+++ Tests/StarknetTests/Devnet/DevnetClientTests.swift
@@ -2,16 +2,19 @@
 import XCTest
 
 final class devnetClientTests: XCTestCase {
-    func testExample() async throws {
-        let test = try makeDevnetClient()
-        try await test.start()
+    var client: DevnetClientProtocol!
 
-        test.close()
+    override func setUp() async throws {
+        client = try makeDevnetClient()
+        try await client.start()
+    }
 
-        // test.prefundAccount(accountAddress: "0x0172d1a003a779c48e66a1b9e591094105d9c48ebc335c44e92faa9197e495cb")
-//        print(try test.deployAccount(name: "test"))
-        // print(test.readAccountDetails(accountName: "test1234"))
-        // You can check that its alive with 'curl http://127.0.0.1:5050/is_alive' in the console
-        // After closing the command above will return an error
+    override func tearDown() async throws {
+        client.close()
+    }
+
+    func testDeployAccount() async throws {
+        let _ = try await client.deployAccount(name: "Account1")
+        let _ = try await client.deployAccount()
     }
 }
